requirements.txt
langchain
pinecone-client
openai
fastapi
uvicorn
llama-index
pydantic
python-dotenv
requests
sentence-transformers